Create index.html
How to Play the Population Definition Dominoes Game
Click the Next button to start the game.
A vocabulary word related to population and its definition will appear on the screen.
Read the word and its definition carefully.
Click Next to see the next word and definition in the sequence.
Continue clicking Next to go through all the dominoes.
When you reach the last domino, the button will change to End. Click it one last time to finish the game.
Use this game to review and memorize key population terms and their meanings.
Good luck and have fun learning!